"Refactor `Calender` and `Therapy` entities to use Lombok annotations for boilerplate code removal"

+++ src/main/resources/db/testdata/0000_test_db.sql
@@ -23,7 +23,7 @@ CREATE TABLE therapy (
 -- Create calender table
 CREATE TABLE calender (
                           id BIGINT PRIMARY KEY AUTO_INCREMENT,
-                          patient_id INT,
+                          patient_id BIGINT,
                           dayOf VARCHAR(20),
                           time VARCHAR(10),
                           free BOOLEAN,
@@ -32,10 +32,10 @@ CREATE TABLE calender (
 
 -- Create meeting table
 CREATE TABLE meeting (
-                         id INT PRIMARY KEY AUTO_INCREMENT,
-                         patient_id INT,
-                         therapy_id INT,
-                         calender_id INT,
+                         id BIGINT PRIMARY KEY AUTO_INCREMENT,
+                         patient_id BIGINT,
+                         therapy_id BIGINT,
+                         calender_id BIGINT,
                          FOREIGN KEY (patient_id) REFERENCES patient(id),
                          FOREIGN KEY (therapy_id) REFERENCES therapy(id),
                          FOREIGN KEY (calender_id) REFERENCES calender(id)
@@ -43,22 +43,22 @@ CREATE TABLE meeting (
 
 -- Create users table
 CREATE TABLE users (
-                       id INT PRIMARY KEY AUTO_INCREMENT,
+                       id BIGINT PRIMARY KEY AUTO_INCREMENT,
                        email VARCHAR(255),
                        password VARCHAR(255)
 );
 
 -- Create user_role table
 CREATE TABLE user_role (
-                           id INT PRIMARY KEY AUTO_INCREMENT,
+                           id BIGINT PRIMARY KEY AUTO_INCREMENT,
                            name VARCHAR(255),
                            description VARCHAR(255)
 );
 
 -- Create user_roles table (junction table for users and user_role)
 CREATE TABLE user_roles (
-                            user_id INT,
-                            role_id INT,
+                            user_id BIGINT,
+                            role_id BIGINT,
                             CONSTRAINT fk_user_roles_user FOREIGN KEY (user_id) REFERENCES users(id),
                             CONSTRAINT fk_user_roles_role FOREIGN KEY (role_id) REFERENCES user_role(id),
                             CONSTRAINT pk_user_roles PRIMARY KEY (user_id, role_id)
@@ -71,8 +71,8 @@ CREATE INDEX fk_user_roles_role_id_index_c ON user_roles (role_id);
 
 -- Create change_password table
 CREATE TABLE change_password (
-                                 id INT PRIMARY KEY AUTO_INCREMENT,
-                                 patient_id INT,
+                                 id BIGINT PRIMARY KEY AUTO_INCREMENT,
+                                 patient_id BIGINT,
                                  old_password VARCHAR(255),
                                  new_password VARCHAR(255),
                                  confirm_new_password VARCHAR(255),
@@ -84,10 +84,10 @@ CREATE TABLE change_password (
 
 -- Create counter table
 CREATE TABLE counter (
-                         id INT PRIMARY KEY AUTO_INCREMENT,
+                         id BIGINT PRIMARY KEY AUTO_INCREMENT,
                          session_id VARCHAR(255),
                          ip VARCHAR(20),
-                         refresh INT,
-                         entry INT,
+                         refresh BIGINT,
+                         entry BIGINT,
                          date DATE
 );
